ai_backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
chromadb
sentence-transformers
python-docx
requests
